The functions module has gotten too big, split it up into several new files.
